refactor (COLOR) moved constructors and html colors in companion object
